requirements.txt
asgiref==3.5.1
dj-database-url==0.5.0
Django==3.2.13
django-crispy-forms==1.14.0
django-heroku==0.3.1
gunicorn==20.1.0
Pillow==9.1.0
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
whitenoise==6.2.0
